+++ src/linter/IcarusLinter.ts
@@ -13,7 +13,8 @@ let standardToArg: Map<string, string> = new Map<string, string>([
 ]);
 
 export default class IcarusLinter extends BaseLinter {
-  private linterDir: string;
+  private configuration: vscode.WorkspaceConfiguration;
+  private linterInstalledPath: string;
   private arguments: string;
   private includePath: string[];
   private standards: Map<string, string>;
@@ -28,27 +29,18 @@ export default class IcarusLinter extends BaseLinter {
   }
 
   private updateConfig() {
-    this.linterDir = <string>vscode.workspace.getConfiguration().get('verilog.linting.path');
-    this.arguments = <string>(
-      vscode.workspace.getConfiguration().get('verilog.linting.iverilog.arguments')
+    this.linterInstalledPath = <string>(
+      vscode.workspace.getConfiguration().get('verilog.linting.path')
     );
-    let path = <string[]>(
-      vscode.workspace.getConfiguration().get('verilog.linting.iverilog.includePath')
-    );
+    this.configuration = vscode.workspace.getConfiguration('verilog.linting.iverilog');
+    this.arguments = <string>this.configuration.get('arguments');
+    let path = <string[]>this.configuration.get('includePath');
     this.includePath = path.map((includePath: string) => this.resolvePath(includePath));
     this.standards = new Map<string, string>([
-      [
-        'verilog',
-        vscode.workspace.getConfiguration().get('verilog.linting.iverilog.verilogHDL.standard'),
-      ],
-      [
-        'systemverilog',
-        vscode.workspace.getConfiguration().get('verilog.linting.iverilog.systemVerilog.standard'),
-      ],
+      ['verilog', this.configuration.get('verilogHDL.standard')],
+      ['systemverilog', this.configuration.get('systemVerilog.standard')],
     ]);
-    this.runAtFileLocation = <boolean>(
-      vscode.workspace.getConfiguration().get('verilog.linting.iverilog.runAtFileLocation')
-    );
+    this.runAtFileLocation = <boolean>this.configuration.get('runAtFileLocation');
   }
 
   // returns absolute path
@@ -60,7 +52,7 @@ export default class IcarusLinter extends BaseLinter {
   }
 
   protected lint(doc: vscode.TextDocument) {
-    this.logger.info('[iverilog-lint] iverilog lint requested');
+    let binPath: string = path.join(this.linterInstalledPath, 'iverilog');
 
     let args: string[] = [];
     args.push('-t null');
@@ -70,7 +62,7 @@ export default class IcarusLinter extends BaseLinter {
     args.push(this.arguments);
     args.push(doc.uri.fsPath);
 
-    let command: string = path.join(this.linterDir, 'iverilog') + ' ' + args.join(' ');
+    let command: string = binPath + ' ' + args.join(' ');
 
     let cwd: string = this.runAtFileLocation
       ? path.dirname(doc.uri.fsPath)
@@ -85,41 +77,49 @@ export default class IcarusLinter extends BaseLinter {
       { cwd: cwd },
       (_error: Error, _stdout: string, stderr: string) => {
         let diagnostics: vscode.Diagnostic[] = [];
-        let lines = stderr.split(/\r?\n/g);
         // Parse output lines
-        lines.forEach((line, _) => {
-          if (line.startsWith(doc.fileName)) {
-            line = line.replace(doc.fileName, '');
-            let terms = line.split(':');
-            let lineNum = parseInt(terms[1].trim()) - 1;
-            if (terms.length == 3) {
-              diagnostics.push({
-                severity: vscode.DiagnosticSeverity.Error,
-                range: new vscode.Range(lineNum, 0, lineNum, Number.MAX_VALUE),
-                message: terms[2].trim(),
-                code: 'iverilog',
-                source: 'iverilog',
-              });
-            } else if (terms.length >= 4) {
-              let sev: vscode.DiagnosticSeverity;
-              if (terms[2].trim() == 'error') {
+        // the message is something like this
+        // /home/ubuntu/project1/module_1.sv:3: syntax error"
+        // /home/ubuntu/project1/property_1.sv:3: error: Invalid module instantiation"
+        stderr.split(/\r?\n/g).forEach((line, _) => {
+          if (!line.startsWith(doc.fileName)) {
+            return;
+          }
+          line = line.replace(doc.fileName, '');
+          let terms = line.split(':');
+          let lineNum = parseInt(terms[1].trim()) - 1;
+          if (terms.length === 3) {
+            diagnostics.push({
+              severity: vscode.DiagnosticSeverity.Error,
+              range: new vscode.Range(lineNum, 0, lineNum, Number.MAX_VALUE),
+              message: terms[2].trim(),
+              code: 'iverilog',
+              source: 'iverilog',
+            });
+          } else if (terms.length >= 4) {
+            let sev: vscode.DiagnosticSeverity;
+            switch (terms[2].trim()) {
+              case 'error':
                 sev = vscode.DiagnosticSeverity.Error;
-              } else if (terms[2].trim() == 'warning') {
+                break;
+              case 'warning':
                 sev = vscode.DiagnosticSeverity.Warning;
-              } else {
+                break;
+              default:
                 sev = vscode.DiagnosticSeverity.Information;
-              }
-              diagnostics.push({
-                severity: sev,
-                range: new vscode.Range(lineNum, 0, lineNum, Number.MAX_VALUE),
-                message: terms[3].trim(),
-                code: 'iverilog',
-                source: 'iverilog',
-              });
             }
+            diagnostics.push({
+              severity: sev,
+              range: new vscode.Range(lineNum, 0, lineNum, Number.MAX_VALUE),
+              message: terms[3].trim(),
+              code: 'iverilog',
+              source: 'Icarus Verilog',
+            });
           }
         });
-        this.logger.info('[iverilog-lint] ' + diagnostics.length + ' errors/warnings returned');
+        if (diagnostics.length > 0) {
+          this.logger.info('[iverilog-lint] ' + diagnostics.length + ' errors/warnings returned');
+        }
         this.diagnosticCollection.set(doc.uri, diagnostics);
       }
     );