+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.x.y] - Unreleased
 
-- Added an experimental option (`verilog.linting.iverilog.includePath`) to specify include directory for iverilog linting.
-  - \[Caution\] I've only tested on Ubuntu on WSL2 platform. If you find any problems about iverilog linter, please let me know.
+- Added options for Icarus Verilog linting.
+  - `verilog.linting.iverilog.includePath` is to specify include directory.
+  - `verilog.linting.iverilog.verilogHDL.standard` is to specify standard rules for Verilog-HDL files.
+  - `verilog.linting.iverilog.systemVerilog.standard` is to specify standard rules for SystemVerilog files.
+  - \[Caution\] I've only tested on Ubuntu on WSL2 platform. If you find any problems about Icarus Verilog linting, please let me know.
 
 ## [1.8.1] - 2022-12-26