+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
   - `verilog.linting.iverilog.verilogHDL.standard` is to specify standard rules for Verilog-HDL files.
   - `verilog.linting.iverilog.systemVerilog.standard` is to specify standard rules for SystemVerilog files.
   - \[Caution\] I've only tested on Ubuntu on WSL2 platform. If you find any problems about Icarus Verilog linting, please let me know.
-- Added an experimental option for Verilator linting.
+- Added an experimental option for Verilator linting. [#380](https://github.com/mshr-h/vscode-verilog-hdl-support/issues/380)
   - `verilog.linting.verilator.includePath` is to specify include directories.
   - \[Caution\] I've only tested on Ubuntu on WSL2 platform. If you find any problems about Verilator linting, please let me know.